+++ packages/ti/osal/soc/j7200/TimerP_default_r5f.c
@@ -224,7 +224,7 @@ void TimerP_updateDefaultInfoTbl(void)
         intBase = CSLR_R5FSS0_CORE0_INTR_TIMER0_INTR_PEND_0;
         for (i = 0U; i < TimerP_numTimerDevices; i++)
         {
-            gDmTimerPInfoTbl[i].baseAddr = CSL_TIMER0_CFG_BASE + (i * 0x10000U);
+            gDmTimerPInfoTbl[i].baseAddr = (uint32_t)CSL_TIMER0_CFG_BASE + (i * 0x10000U);
             gDmTimerPInfoTbl[i].intNum   = intBase + i;
             gDmTimerPInfoTbl[i].eventId  = TIMERP_EVENT_NOT_AVAILABLE;
